refactor: expenseCategory의 _components 폴더 정리
