+++ src/app/home/setting/expenseCategory/_components/ExpenseCategoryContainer/ExpenseCategoryContainer.type.ts
@@ -1,4 +1,4 @@
-import { ExpenseCategory as ExpenseCategoryInterface } from "./Main.type";
+import { ExpenseCategory as ExpenseCategoryInterface } from "../Main/Main.type";
 import { ExpenseCategoryAction } from "@/reducers/expenseCategoryReducer";
 
 export interface ExpenseCategoryContainerProps {
